Replace old login command with new qrlogin oauth command
This removes the old login command file which was querying the ivcap service
directly for a jwt. We now rely on the ivcap service to return a struct
containing oauth provider which we will communicate with.
